Separate `spec` and `status` of orders and clusters
Currently for both orders and clusters all the fields are directly
inside the representation of the type. In the future we will most
probably want to have fields that the user can modify. For example, we
will probably want to have a `number_of_nodes` fields in the cluster
object that can be modified by the user. It will then be difficult to
represent both the value desired by the user and the actual value. To
make that easier this patch splits orders and clusters into `spec` and
`template` sub-objects, similar to what is common practice in
Kubernetes. That way we will be able to hav  a `spec.number_of_nodes`
that the user can modify as well as `status.number_of_nodes` where the
system reports the actual number of nodes.

The template and event types don't need this separation because are read-only.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>