Remove events service
The API for the events service was defined assuming that there will be a
mechanism to reliably deliver changes to clients. But we have decided
that we don't need it. It will be replaced in the future by something
that allows clients to watch specific resources, without a delivery
guarantee.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>