Add cluster node sets
This patch adds to the `Cluster` type a new `node_sets` field to
describe the sets of nodes that are part of the cluster, and to allow
the user to change their size.

For example, a cluster with two node sets, one of them containing hosts
with GPUs and another without could be represented like this:

```json
{
  "id": "123",
  "spec": {
    "node_sets": {
      "compute": {
        "class": "acme_1tb",
        "size": 3
      },
      "gpu": {
        "class": "acme_1tb_h100",
        "size": 1
      }
    }
  },
  "status": {
    "state": "CLUSTER_STATE_READY",
    "node_sets": {
      "compute": {
        "class": "acme_1tb",
        "size": 3
      },
      "gpu": {
        "class": "acme_1tb_h100",
        "size": 1
      }
    }
  }
}
```

The user will be able to scale each node set updating the `size` field
of the `spec`.

The user will not have permission to edit any other field, or to add or
remove node sets.

If the system is not able to resize the node set as requested it will
set the `DEGRADED` condition to `true`, and will indicate the reason in
the `message` field.

In order to provide information about the values of the `class` field of
the node set the API will also have a new `HostClass` type and a
`HostClases` service.

Related: https://github.com/innabox/issues/issues/99
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>