Add `HostPool` and `Host` concepts
This patch adds the `HostPool` and `Host` concepts that will be the
basis for the bare metal host provisioning feature.

The `HostPool` represents a pool of hosts, each containing potentially
multiple sets of hosts. It will contain the fields that allow the user
to modify the pool, or for example to change the number of hosts.

The `Host` represents a host that has been allocated and assigned to a
pool. The `status.host_pool` field will indicate the pool that it
belongs to.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>