Fix names of generated OpenAPI stream results (#46)
The tool that we use to generate OpenAPI specifications generates names
for stream results with spaces, something like this:

```
    Stream result of v1EventsWatchResponse:
      title: Stream result of v1EventsWatchResponse
```

Some OpenAPI tools, like `oq`, don't accept those spaces. To address
that this patch changes the code that generates the OpenAPI schemas so
that it replaces `Stream result of ...` with `...StreamResult`.

Resolves: https://github.com/innabox/fulfillment-api/issues/45
Related: https://github.com/plutov/oq

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>