Add metadata service
This patch adds to the API a new service that provides server metadata.

The first piece of metadata will be the list of trusted token issues.
For example, once it is implemented in the server side it will return
something like this:

```
$ grpcurl -plaintext localhost:8000 metadata.v1.Metadata/Get
{
  "authn": {
    "trustedTokenIssuers": [
      "https://my-issuer.com"
    ]
  }
}
```

This information will be clients (in particular by the CLI) to determine
which OAuth authorization server to use without having to ask the user
to explicitly provde it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>